+++ src/routes/+layout.svelte
@@ -7,10 +7,11 @@
 
 	import { Visualizer } from '$lib'
 	import { backendIP, websocketPort } from '$lib/defines'
-	import { DrawService, Focus, Logs, MeasureTool, XR, XRSettings } from '$lib/plugins'
+	import { AISettings, DrawService, Focus, MeasureTool, XR, XRSettings } from '$lib/plugins'
 
 	import MachineConnectionProvider from './lib/components/MachineConnectionProvider.svelte'
 	import Machines from './lib/components/Machines.svelte'
+	import StandaloneLLMWrapper from './lib/components/StandaloneLLMWrapper.svelte'
 	import {
 		provideConnectionConfigs,
 		useActiveConnectionConfig,
@@ -67,7 +68,10 @@
 			<Visualizer
 				{partID}
 				inputBindingsEnabled={!isMachinesPageOpen}
-				settingsTabs={[{ label: 'AR', component: XRSettings }]}
+				settingsTabs={[
+					{ label: 'AR', component: XRSettings },
+					{ label: 'AI', component: AISettings },
+				]}
 			>
 				{@render children()}
 
@@ -80,6 +84,7 @@
 				<Focus />
 				<MeasureTool />
 				<XR />
+				<StandaloneLLMWrapper />
 			</Visualizer>
 		</MachineConnectionProvider>
 	</ViamAppProvider>