+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # motion-tools
 
+## 1.34.0
+
+### Minor Changes
+
+- 64db2c5: added llm scene builder
+
+## 1.33.2
+
+### Patch Changes
+
+- 054deaf: Fix world state store service race conditions
+
 ## 1.33.1
 
 ### Patch Changes